+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.2.4] - 4/2/26
+### Fixed
+- Proposals that were Submitted previously did not allow editing the contract
+    - This has been fixed; Contracts can be edited until the Proposal is marked as 'Signed'
+        - If another edit to the Contract needs to be made after 'Signed', a Proposal Addendum must be made
+- Exporting a Contract now gives a propoer Print Preview and prints correctly
+
+-----
 ## [2.2.3] - 4/2/26
 ### HOTFIX
 - Fixed unknown draft state issue with Admins